feat: SuccessPanel component, styling
