+++ frontend/src/component/SuccessPanel/index.ts
@@ -1,0 +1,1 @@
+export * from "./SuccessPanel";
