+++ frontend/src/component/index.tsx
@@ -4,3 +4,4 @@ export * from "./Link";
 export * from "./UserCalendar";
 export * from "./WelcomePanel";
 export * from "./SnackBar";
+export * from "./SuccessPanel";
