+++ frontend/src/layout/MainLayout/MainLayout.tsx
@@ -3,7 +3,7 @@ import { Copyright } from "src";
 
 export const MainLayout: FunctionComponent = ({ children }) => {
   return (
-    <div className="w-full md:h-screen flex flex-col justify-center items-center">
+    <div className="w-full min-h-screen flex flex-col justify-center items-center">
       {children}
       <div className="my-8">
         <Copyright />
